requirements.txt
contextlib2==0.6.0.post1
docopt==0.6.2
easywatch==0.0.5
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
pyaml==20.4.0
python-slugify==4.0.1
PyYAML==5.3.1
schema==0.7.2
staticjinja==0.4.0
text-unidecode==1.3
watchdog==1.0.2
